Fix 029_a

+++ java/atcoder/agc029_a.java
@@ -5,8 +5,8 @@ class Main {
         Scanner scanner = new Scanner(System.in);
         String s = scanner.nextLine();
 
-        int result = 0;
-        int whiteCount = 0;
+        long result = 0;
+        long whiteCount = 0;
         for (int i = 0; i < s.length(); i++) {
             char ch = s.charAt(i);
             if (ch == 'W') {
